Move HierarchiclTestBean to demo package.
Moving the Bean to the demo package as it's
used in the demo and other tests can use it from there
also as needed.

This will make the demo build much less brittle due to
changes to the parent project.
