+++ src/test/java/com/vaadin/flow/component/grid/LazyHierarchicalDataProvider.java
@@ -20,6 +20,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Stream;
 
+import com.vaadin.flow.component.grid.demo.HierarchicalTestBean;
 import com.vaadin.flow.data.provider.hierarchy.AbstractBackEndHierarchicalDataProvider;
 import com.vaadin.flow.data.provider.hierarchy.HierarchicalQuery;
