+++ src/test/java/com/vaadin/flow/component/grid/demo/HierarchicalTestBean.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License. 
  */
-package com.vaadin.flow.component.grid;
+package com.vaadin.flow.component.grid.demo;
 
 public class HierarchicalTestBean {
